[WIP]customHookのuseSkills作成
